requirements.txt
frappe>=15.0
erpnext>=15.0
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.8.0
python-docx>=0.8.11
Pillow>=9.0.0
pydub>=0.25.1
SpeechRecognition>=3.10.0
google-cloud-speech>=2.14.0
librosa>=0.9.2
plotly>=5.9.0
chart-studio>=1.1.0
python-jose>=3.3.0
PyJWT>=2.6.0
cryptography>=38.0.0
APScheduler>=3.9.1
celery>=5.2.0
redis>=4.3.0
google-auth>=2.14.0
google-auth-oauthlib>=0.6.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.60.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
Werkzeug>=2.2.0
python-magic>=0.4.25
PyYAML>=6.0
